사용자의 닉네임으로 소유권자인지 확인하는 verifyMemberOwnership 메서드 추가 인자로 nickname을 받습니다 postQuestion 응답 mapper 적용

+++ Server/src/main/java/com/stackoverflow/Server/question/service/QuestionService.java
@@ -5,6 +5,7 @@ import com.stackoverflow.Server.exception.BusinessLogicException;
 import com.stackoverflow.Server.exception.ExceptionCode;
 import com.stackoverflow.Server.member.entity.Member;
 import com.stackoverflow.Server.member.repository.MemberRepository;
+import com.stackoverflow.Server.member.service.MemberService;
 import com.stackoverflow.Server.question.entity.Question;
 import com.stackoverflow.Server.question.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,7 @@ import java.util.Optional;
 @Transactional
 public class QuestionService {
 
+    private final MemberService memberService;
     private final QuestionRepository questionRepository;
     private final MemberRepository memberRepository;
 
@@ -58,8 +60,8 @@ public class QuestionService {
     }
 
     public void removeQuestion(long questionId) {
-
         Question question = findQuestion(questionId);
+        memberService.verifyMemberOwnership(question.getMember().getNickname());
         questionRepository.deleteById(questionId);
     }
